utils:  Update afk function
Apple changed, once again, how the `ScreenSaverEngine` is invoked via
CLI on macOS 10.13 (High Sierra). I've updated the function to check if
the `ScreenSaverEngine.app` is on the Frameworks folder. It it's there,
open it using the full path. Otherwise, rely on `/usr/bin/open` to find
the proper app and run it for us. I hope this is the last time I need to
change this.