tmux:  Update mx
This is a big update to mx. I consider this to be mx 1.0

First, set the window size based on the current terminal when creating
the session. This solve the problem I was having to set my left sidebar
as 15% since when `split-window` runs the session is already the right
size.

Second, add a 'Logs & Notes' window at position 0. There's some magic
going on with `tmux wait-for` to ensure I setup Vim correctly. This will
edit the files `~/_notes.md` and `~/_log.md`

Finally, source the `dotfiles` binary, so I can have access to my log
functions and add a little verbosity to mx, so I'm sure it's not stuck
or anything like that.

Also, just change double quotes to single quotes anywhere double quotes
weren't necessary for interpolation