git: Better private config handling
Keeping my name and e-mail on the gitconfig.private file was causing
trouble. I also removed the include config from ~/.gitconfig, so no
pesky warnings (and breaking things like homebrew or vim-plug because
the git config is messed up).

So I Added a custom config to gitconfig file: dotfiles.decrypted.

By default, that's false, but when the gitconfig.private file is loaded
(and decrypted), that will return true.

The `make private` command will ensure the repository is fully decrypted
and then it will add a include clause to ~/.gitconfig.local (which is
always loaded)