git: Add option to ignore conflicts to hook
If there's a file named `.ignore-conflicts` inside the `.git` dir, skip
the conflict detector hook. The process is not that easy, to ensure that
I *really* want to skip the conflict marker. This came from my necessity
to commit a database dump that had user generated content that would
match the conflict markers for some reason.