utils: Update mx to support forks
I wanted to keep my forks in a separate folder, so I created a `forks`
file inside the `$PROJECTS` folder on my machine and all my forks live
there. So I've adapted `mx` and it's completion to work with this
structure.

`mx` completion now also shows when a mx session is already open